--- Apsitvarkom.Api/Logging/FileLoggerExtensions.cs
@@ -1,11 +1,0 @@
-﻿namespace Apsitvarkom.Api.Logging
-{
-    public static class FileLoggerExtensions
-    {
-        public static ILoggingBuilder AddFile(this ILoggingBuilder factory, string filePath)
-        {
-            factory.AddProvider(new FileLoggerProvider(filePath));
-            return factory;
-        }
-    }
-}
