--- Apsitvarkom.Api/Logging/FileLoggerProvider.cs
@@ -1,21 +1,0 @@
-﻿namespace Apsitvarkom.Api.Logging
-{
-    public class FileLoggerProvider : ILoggerProvider
-    {
-        private string path;
-
-        public FileLoggerProvider(string _path)
-        {
-            path = _path;
-        }
-
-        public ILogger CreateLogger(string categoryName)
-        {
-            return new FileLogger(path);
-        }
-
-        public void Dispose()
-        {
-        }
-    }
-}
