+++ Apsitvarkom.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Apsitvarkom.UnitTests", "Ap
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Apsitvarkom.DataAccess", "Apsitvarkom.DataAccess\Apsitvarkom.DataAccess.csproj", "{655EE043-C640-4F28-B8A4-8CFC73EB86FF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apsitvarkom.Utilities", "Apsitvarkom.Utilities\Apsitvarkom.Utilities.csproj", "{CD97E546-DA91-4A66-B3B0-CD52969D6287}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,10 @@ Global
 		{655EE043-C640-4F28-B8A4-8CFC73EB86FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{655EE043-C640-4F28-B8A4-8CFC73EB86FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{655EE043-C640-4F28-B8A4-8CFC73EB86FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CD97E546-DA91-4A66-B3B0-CD52969D6287}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CD97E546-DA91-4A66-B3B0-CD52969D6287}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CD97E546-DA91-4A66-B3B0-CD52969D6287}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CD97E546-DA91-4A66-B3B0-CD52969D6287}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
